added step entities and dao and usecases

+++ composeApp/src/commonMain/kotlin/data/network/response/task/step/StepDetail.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class StepDetail(
     @SerialName("init_wi")
-    val init_wi : Int,
+    val init_wi : Long,
     @SerialName("activity")
     val acitivities : List<Activity>
 )
