+++ composeApp/src/commonMain/kotlin/domain/models/steps/StepDetailDomain.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 data class StepDetailDomain(
-    val init_wi : Int,
+    val init_wi : Long,
     val acitivities : List<ActivityDomain>
 )
